+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - slightly rework file dialog GUI
 - core/DebugScene : make setupPipelines() private, do not use optional
+- multibody/Visualizer/GUI : display info for robot debug frames
 
 ### Fixed